requirements.txt
requests
bs4
solidpython
Pillow

# Development dependencies
pytest
ruff
